+++ packages/tokens/src/types/baunilha-config.d.ts
@@ -1,0 +1,19 @@
+import { CustomTheme } from '@baunilha/tokens'
+
+type Tokens = 'colors' | 'spacing' | 'typography'
+
+type Output =
+  | string
+  | {
+      scss: string
+      ts: string
+    }
+
+export type BaunilhaConfig = {
+  output: Output
+  unityType?: 'px' | 'rem'
+  fontFamily?: string
+  defaultFontSize?: number
+  include?: Tokens[]
+  theme?: CustomTheme
+}